Change of Plans...
Decided that instead, each assignment would be its own solution. As
such, the Main Solution has been repurposed into a Template (with the
only resulting change really being that the directory containing it was
renamed from 'MainSolution' to '~Template'. Note the ~ is to make it so
the Template folder will always be first when the folders are sorted
alphabetically).